bio-informatics/requirements.txt
Flask==2.3.2
requests==2.31.0
biopython==1.81
matplotlib==3.7.2
ete3==3.1.2
scikit-learn==1.3.1
numpy==1.26.4
FuzzyTM>=0.4.0